Potential fix for code scanning alert no. 42: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to avoid overflow. This can be done by casting one of the operands to uint64_t before performing the multiplication. This way, the multiplication will be done in the uint64_t type, preventing overflow.

Specifically, we will modify the line where the multiplication occurs, casting xsize to uint64_t before multiplying it by ysize.

+++ libwebp-1.0.3/src/enc/backward_references_enc.c
@@ -536,7 +536,7 @@ static int BackwardReferencesLz77Box(int xsize, int ysize,
   int window_offsets_size = 0;
   int window_offsets_new_size = 0;
   uint16_t* const counts_ini =
-      (uint16_t*)WebPSafeMalloc(xsize * ysize, sizeof(*counts_ini));
+      (uint16_t*)WebPSafeMalloc((uint64_t)xsize * ysize, sizeof(*counts_ini));
   int best_offset_prev = -1, best_length_prev = -1;
   if (counts_ini == NULL) return 0;
